requirements.txt
git+https://github.com/django/django@stable/1.7.x#egg=Django
psycopg2
markdown
django-environ
